Relocate solver_style3d to style3d subfolder
