+++ newton/solvers/style3d/__init__.py
@@ -13,20 +13,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from .solver import SolverBase
-from .solver_euler import SemiImplicitSolver
-from .solver_featherstone import FeatherstoneSolver
-from .solver_mujoco import MuJoCoSolver
-from .solver_vbd import VBDSolver
-from .solver_xpbd import XPBDSolver
-from .style3d import Style3DSolver
+from .solver_style3d import Style3DSolver
 
 __all__ = [
-    "FeatherstoneSolver",
-    "MuJoCoSolver",
-    "SemiImplicitSolver",
-    "SolverBase",
     "Style3DSolver",
-    "VBDSolver",
-    "XPBDSolver",
 ]
